feat: Add DatabaseModule and PrismaService
This commit adds the `DatabaseModule` and `PrismaService` files to the project. The `DatabaseModule` is responsible for providing the `PrismaService` as a provider and exporting it for use in other modules. The `PrismaService` is a service that extends the `PrismaClient` from the `@prisma/client` package and handles the connection and disconnection to the database.